rename `erctx` to `erc20tx`
