+++ scenarios/erc20tx/contract/.gitignore
@@ -1,0 +1,3 @@
+*.output.json
+*.bin
+*.abi
