+++ scenarios/erc20tx/contract/TestToken.go
