+++ scenarios/erc20tx/contract/TestToken.sol
